Allow the backfill publisher to look further ahead.
Previously the back-fill publisher was limited to a 2 * batchSize window
when looking ahead of the current cursor for back-fill blocks that can
be enqueued. This change allows the back-fill publisher to look further
ahead, all the way to the tip, so that the back-fill is not stalled by
failing blocks (which prevent the cursor from advancing).