Change batching logic to increase throughput. (#101)
* Change batching logic to increase throughput.

Try to enqueue a new batch whenever the queue has less than batchSize
blocks pending, such that a few pending blocks don't stop the queue from
being refilled. Note that this change also implies that the queue size
will be between 0 and 2*batchSize.

* Allow the backfill publisher to look further ahead.

Previously the back-fill publisher was limited to a 2 * batchSize window
when looking ahead of the current cursor for back-fill blocks that can
be enqueued. This change allows the back-fill publisher to look further
ahead, all the way to the tip, so that the back-fill is not stalled by
failing blocks (which prevent the cursor from advancing).