Allow for blocks to be skipped.
If eksportisto monitor cannot process blocks faster than they're being
produced, allow it to skip to the last block. To put it differently,
eksportisto monitor will only try to process the last block available
when it starts processing, instead of trying to process every single
block received.

This is achieved by adding another go routine that is a fast header
consumer (so as not to stall the HeaderListener) and pushes block
numbers to be processed only when there's a slot available.