Add debug_TraceTransaction timeout option
This option allows the timeout provided to blockchain nodes when making
debug_TraceTransaction calls to be configured.

This update uses a new version of kliento that accepts the timeout
parameter.

Previously the timeout was hardcoded to 50s in kliento, but 50s was not
sufficient for larger transactions.

Going forward, to be safe I suggest 120s for the timeout.