Add transaction tracing timeout option (#118)
This option allows the timeout provided to blockchain nodes when making
debug_TraceTransaction calls to be configured.

This update uses a new version of kliento that accepts the timeout
parameter.

Previously the timeout was hardcoded to 50s in kliento, but 50s was not
sufficient for larger transactions.

Going forward, to be safe I suggest 120s for the timeout, and accordingly
the default timeout is now 120s. 

This change (50s -> 120s) causes the default behaviour of eksportisto to
change, but it should not impact internal clabs users and it looks like there
are very few external users and so impact should be minimal.